requirements.txt
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
cssselect==1.2.0
deprecation==2.1.0
Flask==2.3.2
functions-framework==3.3.0
google-api-core==2.11.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-cloud-secret-manager==2.16.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
gspread==5.9.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
loguru==0.7.0
lxml==4.9.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.1
parsel==1.8.1
proto-plus==1.22.2
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.15
w3lib==2.1.1
watchdog==3.0.0
Werkzeug==2.3.4
